+++ src/main/java/micycle/pgs/PGS_Morphology.java
@@ -151,7 +151,7 @@ public final class PGS_Morphology {
 		try {
 			return toPShape(JTS.smooth(fromPShape(shape), 1 - fit));
 		} catch (Exception e) {
-			return new PShape();
+			return shape;
 		}
 
 	}
@@ -179,7 +179,7 @@ public final class PGS_Morphology {
 			return toPShape(GaussianLineSmoothing.get(l, Math.max(sigma, 1)));
 		}
 		System.err.println(g.getGeometryType() + " are not supported for this method (yet).");
-		return new PShape(PConstants.GROUP);
+		return shape;
 	}
 
 	/**
@@ -327,28 +327,46 @@ public final class PGS_Morphology {
 	 */
 	public static PShape fieldWarp(PShape shape, double magnitude, double noiseScale, double time, boolean densify, int noiseSeed) {
 		float scale = (float) noiseScale * 500f;
-		final List<PVector> coords;
+		final boolean pointsShape = shape.getKind() == PConstants.POINTS;
 
-		if (densify) {
+		final PShape copy;
+		if (densify && !pointsShape) {
 			final Densifier d = new Densifier(fromPShape(shape));
 			d.setDistanceTolerance(1);
 			d.setValidate(false);
-			coords = PGS_Conversion.toPVector(toPShape(d.getResultGeometry()));
+			copy = toPShape(d.getResultGeometry());
 		} else {
-			coords = PGS_Conversion.toPVector(shape);
+			copy = toPShape(fromPShape(shape));
 		}
 
 		final UniformNoise noise = new UniformNoise(noiseSeed);
 
-		coords.forEach(coord -> {
-//			float dx = noise.uniformNoise(coord.x / scale, coord.y / scale, time) - 0.5f;
-			float dx = noise.uniformNoise(coord.x / scale, coord.y / scale + time) - 0.5f;
-//			float dy = noise.uniformNoise(coord.x / scale, coord.y / scale, 100 + time) - 0.5f;
-			float dy = noise.uniformNoise(coord.x / scale + (101 + time), coord.y / scale + (101 + time)) - 0.5f;
-			coord.add(dx * (float) magnitude * 2, dy * (float) magnitude * 2);
-		});
-		
-		return toPShape(GeometryFixer.fix(fromPShape(PGS_Conversion.fromPVector(coords))));
+		if (copy.getChildCount() == 0) {
+			// setVertex() will act on group shapes, so treat a single shape as group of 1
+			copy.addChild(copy);
+		}
+
+		for (PShape child : copy.getChildren()) {
+			for (int i = 0; i < child.getVertexCount(); i++) {
+				final PVector coord = child.getVertex(i);
+				float dx = noise.uniformNoise(coord.x / scale, coord.y / scale + time) - 0.5f;
+				float dy = noise.uniformNoise(coord.x / scale + (101 + time), coord.y / scale + (101 + time)) - 0.5f;
+				child.setVertex(i, coord.x + (dx * (float) magnitude * 2), coord.y + (dy * (float) magnitude * 2));
+			}
+		}
+
+		if (pointsShape) {
+			return copy;
+		} else {
+
+			if (copy.getChildCount() == 1) {
+				return toPShape(GeometryFixer.fix(fromPShape(copy.getChild(0))));
+			} else {
+				// don't apply geometryFixer to GROUP shape, since fixing a multigeometry
+				// appears to merge shapes. TODO apply .fix() to shapes individually
+				return copy;
+			}
+		}
 	}
 
 }